% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_duplicates}
\alias{check_duplicates}
\title{check_duplicates}
\usage{
check_duplicates(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
A dataframe of duplicate rows (or message if none found).
}
\description{
Find duplicate rows in a dataframe.
}
\examples{
check_duplicates(df)
}
