% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/helpers.R
\name{check_outliers}
\alias{check_outliers}
\title{check_outliers}
\usage{
check_outliers(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
A dataframe with count of outliers per numeric column.
}
\description{
Identify numeric columns with values beyond 1.5x IQR.
}
\examples{
check_outliers(df)
}
