% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/helpers.R
\name{describe_data}
\alias{describe_data}
\title{describe_data}
\usage{
describe_data(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
A list with numeric and categorical summaries.
}
\description{
Provide summary statistics for numeric and categorical columns.
}
\examples{
describe_data(df)
}
