% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/import_sql.R
\name{import_sql}
\alias{import_sql}
\title{Import Sql}
\usage{
import_sql(sql_file, server, database, uid, pwd, driver = "SQL Server")
}
\arguments{
\item{...}{Parameters passed to \code{import_sql}.}
}
\value{
Output from \code{import_sql}.
}
\description{
Import Sql function from EMSqiTools package.
}
