% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/plot_imr_chart.R
\name{plot_imr_chart}
\alias{plot_imr_chart}
\title{Plot Imr Chart}
\usage{
plot_imr_chart(
  df,
  date_col,
  id_col,
  value_col,
  den_condition = "TRUE",
  time_unit = c("week", "month", "quarter"),
  name = "I-MR Chart",
  annotations = NULL
)
}
\arguments{
\item{...}{Parameters passed to \code{plot_imr_chart}.}
}
\value{
Output from \code{plot_imr_chart}.
}
\description{
Plot Imr Chart function from EMSqiTools package.
}
