% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{plot_table}
\usage{
plot_table(
  df,
  date_col,
  id_col,
  num_condition,
  den_condition = "TRUE",
  time_unit = c("week", "month", "quarter")
)
}
\arguments{
\item{df}{Dataframe.}

\item{date_col}{Name of the date column (string).}

\item{id_col}{Unique identifier column (string).}

\item{num_condition}{Filter condition for numerator (string, R expression).}

\item{den_condition}{Filter condition for denominator (default 'TRUE').}

\item{time_unit}{Time unit for aggregation ('week', 'month', 'quarter').}
}
\value{
A summary dataframe.
}
\description{
Create a summary table with numerator, denominator, percent, and formatted percent.
}
\examples{
table <- plot_table(df, "Incident Date", "pcr", "c02 < 45", "TRUE", "week")
}
