% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/helpers.R
\name{profile_data}
\alias{profile_data}
\title{profile_data}
\usage{
profile_data(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
An HTML report opened in the browser.
}
\description{
Generate an interactive data profile report using DataExplorer.
}
\examples{
profile_data(df)
}
