% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qi_sql.R
\name{qi_sql}
\alias{qi_sql}
\title{qi_sql}
\usage{
qi_sql(sql_file, server, database, uid, pwd, driver = "SQL Server")
}
\arguments{
\item{sql_file}{Path to the .sql file containing the query.}

\item{server}{Database server name or address.}

\item{database}{Database name.}

\item{uid}{User ID (username).}

\item{pwd}{Password.}

\item{driver}{ODBC driver name (default: "SQL Server").}
}
\value{
A dataframe with the query result.
}
\description{
Run a SQL query from a .sql file against a database server.
}
\examples{
df <- qi_sql("query.sql", server = "my_server", database = "my_db", uid = "username", pwd = "password")
}
