% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Summary Table}
\usage{
summary_table(
  df,
  date_col,
  numerator_col = NULL,
  denominator_col = NULL,
  numerator_value = NULL,
  plot_type = c("proportion", "count", "average"),
  time_col = NULL,
  start_date,
  end_date
)
}
\arguments{
\item{...}{Parameters passed to \code{summary_table}.}
}
\value{
Output from \code{summary_table}.
}
\description{
Summary Table function from EMSqiTools package.
}
